+++ apps/backend/src/libs/hooks/hooks.ts
@@ -1,1 +1,2 @@
+export { checkHasPermission } from "./check-permission/check-permission.js";
 export { setRateLimit } from "./set-rate-limit/set-rate-limit.hook.js";