+++ apps/backend/src/libs/modules/aws/libs/enums/aws-file.exeption-message.ts
@@ -1,0 +1,5 @@
+const AwsExceptionMessage = {
+	FILE_DOWNLOADING_ERROR: "The problem occured while downloading the file",
+};
+
+export { AwsExceptionMessage };