+++ packages/shared/src/index.ts
@@ -58,8 +58,13 @@ export {
 	NotificationType,
 } from "./modules/notifications/notifications.js";
 export {
+	type PermissionItemDto,
+	PermissionKey,
+} from "./modules/permission/permission.js";
+export {
 	pointOfInterestCreateValidationSchema,
 	pointOfInterestUpdateValidationSchema,
+	PointsOfInterestApiPath,
 	type PointsOfInterestRequestDto,
 	type PointsOfInterestResponseDto,
 	type PointsOfInterestSearchQuery,